fix(wb): re-employ memoization and upgrade dependencies
