+++ package.json
@@ -25,8 +25,8 @@
   "prettier": "@willbooster/prettier-config",
   "devDependencies": {
     "@qiwi/multi-semantic-release": "7.1.1",
-    "@types/eslint": "8.44.6",
-    "@typescript-eslint/parser": "6.9.1",
+    "@types/eslint": "8.44.7",
+    "@typescript-eslint/parser": "6.10.0",
     "@willbooster/prettier-config": "9.1.2",
     "conventional-changelog-conventionalcommits": "6.1.0",
     "eslint": "8.53.0",
