+++ packages/shared-lib-blitz-next/package.json
@@ -36,10 +36,10 @@
   },
   "prettier": "@willbooster/prettier-config",
   "devDependencies": {
-    "@types/eslint": "8.44.6",
+    "@types/eslint": "8.44.7",
     "@types/micromatch": "4.0.4",
-    "@typescript-eslint/eslint-plugin": "6.9.1",
-    "@typescript-eslint/parser": "6.9.1",
+    "@typescript-eslint/eslint-plugin": "6.10.0",
+    "@typescript-eslint/parser": "6.10.0",
     "@willbooster/eslint-config-ts": "10.5.1",
     "@willbooster/prettier-config": "9.1.2",
     "blitz": "2.0.0-beta.35",
