+++ packages/shared-lib-node/package.json
@@ -40,11 +40,11 @@
     "tree-kill": "1.2.2"
   },
   "devDependencies": {
-    "@types/eslint": "8.44.6",
+    "@types/eslint": "8.44.7",
     "@types/micromatch": "4.0.4",
     "@types/node": "20.8.10",
-    "@typescript-eslint/eslint-plugin": "6.9.1",
-    "@typescript-eslint/parser": "6.9.1",
+    "@typescript-eslint/eslint-plugin": "6.10.0",
+    "@typescript-eslint/parser": "6.10.0",
     "@willbooster/eslint-config-ts": "10.5.1",
     "@willbooster/prettier-config": "9.1.2",
     "build-ts": "11.0.11",
