+++ packages/shared-lib-react/package.json
@@ -35,21 +35,21 @@
   "devDependencies": {
     "@babel/core": "7.23.2",
     "@mdx-js/react": "3.0.0",
-    "@storybook/addon-actions": "7.5.2",
-    "@storybook/addon-docs": "7.5.2",
-    "@storybook/addon-essentials": "7.5.2",
-    "@storybook/addon-interactions": "7.5.2",
-    "@storybook/addon-links": "7.5.2",
+    "@storybook/addon-actions": "7.5.3",
+    "@storybook/addon-docs": "7.5.3",
+    "@storybook/addon-essentials": "7.5.3",
+    "@storybook/addon-interactions": "7.5.3",
+    "@storybook/addon-links": "7.5.3",
     "@storybook/builder-webpack4": "6.5.16",
     "@storybook/manager-webpack4": "6.5.16",
-    "@storybook/react": "7.5.2",
+    "@storybook/react": "7.5.3",
     "@storybook/testing-library": "0.2.2",
-    "@types/eslint": "8.44.6",
+    "@types/eslint": "8.44.7",
     "@types/micromatch": "4.0.4",
-    "@types/react": "18.2.34",
+    "@types/react": "18.2.36",
     "@types/react-dom": "18.2.14",
-    "@typescript-eslint/eslint-plugin": "6.9.1",
-    "@typescript-eslint/parser": "6.9.1",
+    "@typescript-eslint/eslint-plugin": "6.10.0",
+    "@typescript-eslint/parser": "6.10.0",
     "@willbooster/eslint-config-ts-react": "10.1.9",
     "@willbooster/prettier-config": "9.1.2",
     "babel-loader": "9.1.3",
