+++ packages/wb/package.json
@@ -30,16 +30,16 @@
     "tree-kill": "1.2.2"
   },
   "devDependencies": {
-    "@types/eslint": "8.44.6",
-    "@types/kill-port": "2.0.2",
+    "@types/eslint": "8.44.7",
+    "@types/kill-port": "2.0.3",
     "@types/micromatch": "4.0.4",
     "@types/node": "20.8.10",
-    "@types/yargs": "17.0.29",
-    "@typescript-eslint/eslint-plugin": "6.9.1",
-    "@typescript-eslint/parser": "6.9.1",
+    "@types/yargs": "17.0.30",
+    "@typescript-eslint/eslint-plugin": "6.10.0",
+    "@typescript-eslint/parser": "6.10.0",
     "@willbooster/eslint-config-ts": "10.5.1",
     "@willbooster/prettier-config": "9.1.2",
-    "at-decorators": "1.2.2",
+    "at-decorators": "1.3.1",
     "build-ts": "11.0.11",
     "eslint": "8.53.0",
     "eslint-config-prettier": "9.0.0",
